Move tests into seperate test file
