+++ crates/layout/src/inline/mod.rs
@@ -10,6 +10,9 @@ mod textarea;
 mod tokens;
 mod types;
 
+#[cfg(test)]
+mod tests;
+
 pub use button::button_label_from_layout;
 pub(crate) use dom_attrs::get_attr;
 pub use engine::layout_inline_for_paint;
